add some reflection helper stuff
